requirements.txt
stoq-framework
lief